mcp-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx>=0.28.1
ddgs>=9.8.0
googlesearch-python>=1.3.0
